pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bytez>=2.0.2",
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "groq>=0.33.0",
    "gunicorn>=23.0.0",
    "numpy>=2.3.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]
